+++ feagen/decorators/__init__.py
@@ -1,7 +1,8 @@
-from .decorators import (
-    can_skip,
+from .data_generation import (
     require_intermediate_data,
     will_generate,
+)
+from .data_type import (
     intermediate_data,
     features,
 )
